+++ src/main/resources/application.yml
@@ -4,7 +4,7 @@ spring:
 
   profiles:
     group:
-      local: jwt
+      local: jwt, security
 
   datasource:
     url: jdbc:h2:mem:testdb
